+++ gemini/ped.py
@@ -1,3 +1,9 @@
+"""Parse input PED files for loading into GEMINI sample table.
+
+Tries to parse as tab-delimited if files contain tabs, otherwise splits
+on whitespace.
+"""
+
 default_ped_fields = ["family_id", "name", "paternal_id", "maternal_id",
                       "sex", "phenotype"]
 missing_member = set(["None", None, "0", "-9"])
@@ -10,10 +16,13 @@ def get_ped_fields(ped_file):
         possible_header = in_handle.readline()
 
     if possible_header.startswith("#"):
-        header = possible_header.replace("#", "").split()
+        if possible_header.count("\t") > 1:
+            header = possible_header.replace("#", "").rstrip().split("\t")
+        else:
+            header = possible_header.replace("#", "").split()
         # rename the standard fields to a common name
         header = default_ped_fields + header[len(default_ped_fields):]
-        return possible_header.replace("#", "").split()
+        return header
     else:
         return default_ped_fields
 
@@ -22,7 +31,8 @@ def load_ped_file(ped_file):
     for line in open(ped_file, 'r'):
         if line.startswith("#") or len(line) == 0:
             continue
-        fields = _fix_ped_family_fields(line.split())
+        parts = line.rstrip().split("\t") if line.count("\t") > 1 else line.split()
+        fields = _fix_ped_family_fields(parts)
         ped_dict[fields[1]] = fields
     return ped_dict